SSO-49 Added support for customized header and navbar color per district
A field themeColor was added to the groups table in the database and 
this is read and passed to the layout when the user’s tab is displayed. 
The layout.phtml uses JavaScript to calculate a highlight color and 
whether to display the text in black or white based on the brightness of 
the theme color.
A database update is required when deploying this change. The file is 
data/updates/sso-48-district-logo.sql