SSO-55: partial completion of tab layout manager
A DB update is needed (see data/updates/sso-55-app-order.sql) to support 
an additional field to track the display order of apps on a tab.  
Extended the API to support listing related/pivot records such as 
tabApps. Updated javascript on dashboard index.phtml to support the tab 
layout manager. The tab layout manager is not complete yet.